fix(core): resolve PR #626 review minors
- Extract shared token-parity resolvers (readJson, collectColorLeaves,
  parseToOklch, baseLeaves, primitiveColors) into token-parity-utils.ts so
  tone-parity and static-parity can't drift (SDE2-2).
- Guard the static-parity resolver's hidden dependency: pin that the 75/150
  near-white primitives stay off the perceptual grid, with an explanatory
  comment on primitiveLightness (Arch-3).
- Reconcile the spec doc: list the four extra adopted components
  (checkbox, radio-group, date-picker, alert close) in the scope guard
  (Arch-1); correct popover-hover to 0.946/stone.100 in the ladder table and
  step block to match shipped tokens (Arch-2).
